feat: Refactored beer endpoints - split "beer" into "getRecentBeers" and "register". Removed the need for auth to fetch lates beers
